added a readme and included the needed images and screenshots
