+++ Objects_and_Classes/.gitignore
@@ -1,0 +1,1 @@
+app/
